[acimense] RecyclerView works with children queue for flipping and layout/UI/colors have been improved to look nicer.
Fixed a bug where a button from the FlipCoinActivity was showing up in the fragment. Added a Cancel button (unresponsive) which will allow the user to go back to the FlipCoinActivity. TODO will add another button to allow the user to choose "No Selection" for child. The items in the recyclerView are unresponsive, will make these respond to clicks.