[acimense] Added a (unresponsive) button to allow user to select no children for flipping a coin and fixed some fragment issues
Added fragment to back stack from FlipCoinActivity, so I can pop it off the stack by pressing the cancel button. I changed the layout of the FlipCoinActivity, the constraint layout I was using for the bottom two buttons were shifting to the left after returning from the fragment. Had something to do with the fact I was using 0dp as the width I suspect.